Finish adding support for 2.1 billion items

+++ Comet-API/src/main/java/com/cometproject/api/game/players/IPlayer.java
@@ -69,9 +69,9 @@ public interface IPlayer {
 
     boolean isTeleporting();
 
-    int getTeleportId();
+    long getTeleportId();
 
-    void setTeleportId(int teleportId);
+    void setTeleportId(long teleportId);
 
     long getRoomLastMessageTime();
