+++ Comet-Server/src/main/java/com/cometproject/server/game/rooms/objects/items/RoomItemFactory.java
@@ -188,7 +188,7 @@ public class RoomItemFactory {
         } else {
             if (itemDefinitionMap.containsKey(def.getInteraction())) {
                 try {
-                    floorItem = itemDefinitionMap.get(def.getInteraction()).getConstructor(int.class, int.class, Room.class, int.class, int.class, int.class, double.class, int.class, String.class)
+                    floorItem = itemDefinitionMap.get(def.getInteraction()).getConstructor(long.class, int.class, Room.class, int.class, int.class, int.class, double.class, int.class, String.class)
                             .newInstance(id, baseId, room, ownerId, x, y, height, rot, data);
                 } catch (Exception e) {
                     log.warn("Failed to create instance for item: " + id + ", type: " + def.getInteraction(), e);
