+++ Comet-Server/src/main/java/com/cometproject/server/game/rooms/objects/items/types/floor/wired/actions/WiredActionTeleportPlayer.java
@@ -57,7 +57,7 @@ public class WiredActionTeleportPlayer extends WiredActionItem {
             return;
         }
 
-        Integer itemId = WiredUtil.getRandomElement(this.getWiredData().getSelectedIds());
+        Long itemId = WiredUtil.getRandomElement(this.getWiredData().getSelectedIds());
 
         if (itemId == null) {
             this.entity = null;
