+++ Comet-Server/src/main/java/com/cometproject/server/game/rooms/objects/items/types/floor/wired/data/WiredActionItemData.java
@@ -9,7 +9,7 @@ import java.util.Map;
 public class WiredActionItemData extends WiredItemData {
     private int delay;
 
-    public WiredActionItemData(int selectionType, List<Integer> selectedIds, String text, Map<Integer, Integer> params, Map<Integer, WiredItemSnapshot> snapshots, int delay) {
+    public WiredActionItemData(int selectionType, List<Long> selectedIds, String text, Map<Integer, Integer> params, Map<Long, WiredItemSnapshot> snapshots, int delay) {
         super(selectionType, selectedIds, text, params, snapshots);
         this.delay = delay;
     }
