+++ Comet-Server/src/main/java/com/cometproject/server/storage/queries/items/TeleporterDao.java
@@ -10,7 +10,7 @@ import java.sql.SQLException;
 
 public class TeleporterDao {
 
-    public static int getPairId(int id) {
+    public static long getPairId(long id) {
         Connection sqlConnection = null;
         PreparedStatement preparedStatement = null;
         ResultSet resultSet = null;
@@ -19,7 +19,7 @@ public class TeleporterDao {
             sqlConnection = SqlHelper.getConnection();
 
             preparedStatement = SqlHelper.prepare("SELECT * FROM items_teles WHERE id_one = ? LIMIT 1;", sqlConnection);
-            preparedStatement.setInt(1, id);
+            preparedStatement.setLong(1, id);
 
             resultSet = preparedStatement.executeQuery();
 
@@ -37,7 +37,7 @@ public class TeleporterDao {
         return 0;
     }
 
-    public static void savePair(int item1, int item2) {
+    public static void savePair(long item1, long item2) {
         Connection sqlConnection = null;
         PreparedStatement preparedStatement = null;
 
@@ -45,13 +45,13 @@ public class TeleporterDao {
             sqlConnection = SqlHelper.getConnection();
 
             preparedStatement = SqlHelper.prepare("INSERT into items_teles (id_one, id_two) VALUES(?, ?);", sqlConnection);
-            preparedStatement.setInt(1, item1);
-            preparedStatement.setInt(2, item2);
+            preparedStatement.setLong(1, item1);
+            preparedStatement.setLong(2, item2);
 
             preparedStatement.addBatch();
 
-            preparedStatement.setInt(1, item2);
-            preparedStatement.setInt(2, item1);
+            preparedStatement.setLong(1, item2);
+            preparedStatement.setLong(2, item1);
 
             preparedStatement.addBatch();
